Adds dummy-coded variables to data frame.

This function adds dummy-coded variables to a data frame based on a vector of column names.

dummyCode(df, vars)

Arguments

df
A data frame.

vars
The variables in the data frame for which you want to create new dummy coded variables.

Value

A a dataframe with new dummy-coded variables added.

Examples

# Create an example data frame:
df <- data.frame(
    animal   = c("dog", "goldfish", "bird", "dog", "goldfish"),
    numLegs  = c(4, 0, 2, 4, 0),
    lifeSpan = c(10, 10, 5, 10, 10))

# Create dummy coded variables for the variables "animal" and "numLegs":
df_dummy <- dummyCode(df, vars = c("animal", "numLegs"))
df_dummy

#>     animal numLegs lifeSpan animalbird animaldog animalgoldfish numLegs
#> 1      dog       4       10          0         1              0       4
#> 2 goldfish       0       10          0         0              1       0
#> 3     bird       2        5          1         0              0       2
#> 4      dog       4       10          0         1              0       4
#> 5 goldfish       0       10          0         0              1       0