Predict probabilities and / or outcomes

This method is used for computing predicted probabilities and / or outcomes for either the data used for model estimation or a new data set consisting of a single or multiple sets of alternatives.

# S3 method for logitr
predict(
  object,
  newdata = NULL,
  obsID = NULL,
  price = NULL,
  type = "prob",
  returnData = FALSE,
  ci = NULL,
  numDrawsCI = 10^3,
  ...
)

Arguments

object
is an object of class logitr (a model estimated using the 'logitr()` function).

newdata
a data.frame. Each row is an alternative and each column an attribute corresponding to parameter names in the estimated model. Defaults to NULL, in which case predictions are made on the original data used to estimate the model.

obsID
The name of the column that identifies each set of alternatives in the data. Required if newdata != NULL. Defaults to NULL, in which case the value for obsID from the data in object is used.

price
The name of the column that identifies the price variable. Required if the object is a WTP space model and if newdata != NULL. Defaults to NULL.

type
A character vector defining what to predict: prob for probabilities, outcomes for outcomes. If you want both outputs, use c("prob", "outcome"). Outcomes are predicted randomly according to the predicted probabilities. Defaults to "prob".

returnData
If TRUE the data is also returned, otherwise only the predicted values ("prob" and / or "outcome") are returned. Defaults to FALSE.

ci
If a confidence interval (CI) for the predicted probabilities is desired, set ci to a number between 0 and 1 to define the CI sensitivity. For example, ci = 0.95 will return a 95% CI. Defaults to NULL, in which case no CI is computed.

numDrawsCI
The number of draws to use in simulating uncertainty for the computed CI. Defaults to 10^3.

...
further arguments.

Value

A data frame of predicted probabilities and / or outcomes.

Examples

library(logitr)

# Estimate a preference space model
mnl_pref <- logitr(
  data    = yogurt,
  outcome = "choice",
  obsID   = "obsID",
  pars    = c("price", "feat", "brand")
)

#> Running Model...

#> Done!

# Predict probabilities and / or outcomes

# Predict probabilities for each alternative in the model data
probs <- predict(mnl_pref)
head(probs)

#>   obsID predicted_prob
#> 1     1     0.41802407
#> 2     1     0.02118240
#> 3     1     0.23691737
#> 4     1     0.32387615
#> 5     2     0.26643822
#> 6     2     0.02255486

# Create a set of alternatives for which to make predictions.
# Each row is an alternative and each column an attribute.
data <- subset(
    yogurt, obsID %in% c(42, 13),
    select = c('obsID', 'alt', 'price', 'feat', 'brand'))
data

#> # A tibble: 8 × 5
#>   obsID   alt price  feat brand  
#>   <int> <int> <dbl> <dbl> <chr>  
#> 1    13     1  8.1      0 dannon 
#> 2    13     2  5.00     0 hiland 
#> 3    13     3  8.60     0 weight 
#> 4    13     4 10.8      0 yoplait
#> 5    42     1  6.30     0 dannon 
#> 6    42     2  6.10     1 hiland 
#> 7    42     3  7.90     0 weight 
#> 8    42     4 11.5      0 yoplait

# Predict probabilities using the estimated model
predict(mnl_pref, newdata = data, obsID = "obsID")

#>   obsID predicted_prob
#> 1    13     0.43685145
#> 2    13     0.03312986
#> 3    13     0.19155548
#> 4    13     0.33846321
#> 5    42     0.60764778
#> 6    42     0.02602007
#> 7    42     0.17803313
#> 8    42     0.18829902

# Predict outcomes
predict(mnl_pref, newdata = data, obsID = "obsID", type = "outcome")

#>   obsID predicted_outcome
#> 1    13                 1
#> 2    13                 0
#> 3    13                 0
#> 4    13                 0
#> 5    42                 1
#> 6    42                 0
#> 7    42                 0
#> 8    42                 0

# Predict outcomes and probabilities
predict(mnl_pref, newdata = data, obsID = "obsID", type = c("prob", "outcome"))

#>   obsID predicted_prob predicted_outcome
#> 1    13     0.43685145                 1
#> 2    13     0.03312986                 0
#> 3    13     0.19155548                 0
#> 4    13     0.33846321                 0
#> 5    42     0.60764778                 1
#> 6    42     0.02602007                 0
#> 7    42     0.17803313                 0
#> 8    42     0.18829902                 0