Predict choices

Returns the expected choices for a set of one or more alternatives based on the results from an estimated model.

predictChoices(model, alts, altID, obsID = NULL)

Arguments

model
The output of a model estimated model using the logitr() function. Include if you want to compare true choices from actual observations (e.g. hold outs) to the predicted choices.

alts
A data frame of a set of alternatives for which to predict choices. Each row is an alternative and each column an attribute corresponding to parameter names in the estimated model.

altID
The name of the column that identifies each alternative in each set of alternatives.

obsID
The name of the column that identifies each set of alternatives. Required if predicting results for more than one set of alternatives. Defaults to NULL (for a single set of alternatives).

Value

A data frame with the predicted choices for each alternative in alts.

Examples

library(logitr)

# Estimate a preference space model
mnl_pref <- logitr(
  data   = yogurt,
  choice = "choice",
  obsID  = "obsID",
  pars   = c("price", "feat", "brand")
)

#> Running Model...

#> Done!

# You can predict choices for any set of alternative, such as hold out
# samples or within-sample. For this example, choices will be predicted for
# the full yogurt data set, which was used to estimate the model.

# Predict choices using the estimated preference space MNL model
choices <- predictChoices(
  model = mnl_pref,
  alts  = yogurt,
  altID = "alt",
  obsID = "obsID"
)

head(choices)

#>     id obsID alt choice price feat   brand dannon hiland weight yoplait
#> 1.1  1     1   1      0   8.1    0  dannon      1      0      0       0
#> 1.2  1     1   2      0   6.1    0  hiland      0      1      0       0
#> 1.3  1     1   3      1   7.9    0  weight      0      0      1       0
#> 1.4  1     1   4      0  10.8    0 yoplait      0      0      0       1
#> 2.5  1     2   1      1   9.8    0  dannon      1      0      0       0
#> 2.6  1     2   2      0   6.4    0  hiland      0      1      0       0
#>     choice_predict
#> 1.1              0
#> 1.2              0
#> 1.3              0
#> 1.4              1
#> 2.5              0
#> 2.6              0

# Compute the accuracy
chosen <-  subset(choices, choice == 1)
chosen$correct <- chosen$choice == chosen$choice_predict
sum(chosen$correct) / nrow(chosen) # % correctly predicted

#> [1] 0.3884743