Simulate expected market shares

Returns the expected shares of a specific set of alternatives based on an estimated model.

simulateShares(model, alts, priceName, alpha)

Arguments

model
The output of a model estimated using the logitr() function.

alts
A data frame of a set of alternatives for which to simulate shares. Each row is an alternative and each column an attribute corresponding to parameter names in the estimated model.

priceName
The name of the parameter that identifies price.

alpha
The sensitivity of the computed confidence interval, e.g. a 90% CI is obtained with alpha = 0.05. Defaults to alpha = 0.025.

Value

A data frame with the estimated shares for each alternative in alts.

Examples

# Run a MNL model in the Preference Space:
library(logitr)

mnl_pref = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'))

#> Running Model
#> Done!

# Create a set of alternatives for which to simulate shares:
market = subset(yogurt, obsID == 42,
       select=c('feat', 'price', 'dannon', 'hiland', 'yoplait'))
row.names(market) = c('dannon', 'hiland', 'weight', 'yoplait')
market

#>         feat price dannon hiland yoplait
#> dannon     0   6.3      1      0       0
#> hiland     1   6.1      0      1       0
#> weight     0   7.9      0      0       0
#> yoplait    0  11.5      0      0       1

# Run the simulation using the estimated preference space MNL model:
simulateShares(mnl_pref, market)

#>              share_mean share_low share_high
#> Alt: dannon  0.60767171 0.5468190 0.65930218
#> Alt: hiland  0.02601787 0.0186844 0.03641504
#> Alt: weight  0.17802338 0.1630885 0.19251799
#> Alt: yoplait 0.18828704 0.1387648 0.25029720