The main function for estimating logit models

Use this function to estimate multinomial (MNL) and mixed logit (MXL) models with "Preference" space or "Willingness-to-pay" (WTP) space utility parameterizations. The function includes an option to run a multistart optimization loop with random starting points in each iteration, which is useful for non-convex problems like MXL models or models with WTP space utility parameterizations. The main optimization loop uses the nloptr() function to minimize the negative log-likelihood function.

logitr(
  data,
  choice,
  obsID,
  pars,
  price = NULL,
  randPars = NULL,
  randPrice = NULL,
  modelSpace = "pref",
  weights = NULL,
  cluster = NULL,
  robust = FALSE,
  numMultiStarts = 1,
  useAnalyticGrad = TRUE,
  scaleInputs = TRUE,
  startParBounds = c(-1, 1),
  standardDraws = NULL,
  numDraws = 50,
  startVals = NULL,
  options = list(print_level = 0, xtol_rel = 1e-06, xtol_abs = 1e-06, ftol_rel = 1e-06,
    ftol_abs = 1e-06, maxeval = 1000, algorithm = "NLOPT_LD_LBFGS")
)

Arguments

data
The choice data, formatted as a data.frame object.

choice
The name of the column that identifies the choice variable.

obsID
The name of the column that identifies each choice observation.

pars
The names of the parameters to be estimated in the model. Must be the same as the column names in the data argument. For WTP space models, do not include price in pars.

price
The name of the column that identifies the price variable. Required for WTP space models. Defaults to NULL.

randPars
A named vector whose names are the random parameters and values the distribution: 'n' for normal or 'ln' for log-normal. Defaults to NULL.

randPrice
The random distribution for the price parameter: 'n' for normal or 'ln' for log-normal. Only used for WTP space MXL models. Defaults to NULL.

modelSpace
Set to 'wtp' for WTP space models. Defaults to "pref".

weights
The name of the column that identifies the weights to be used in model estimation. Defaults to NULL.

cluster
The name of the column that identifies the cluster groups to be used in model estimation. Optional. Defaults to NULL.

robust
Determines whether or not a robust covariance matrix is estimated. Defaults to FALSE. Specification of a cluster or weights will override the user setting and set this to `TRUE' (a warning will be displayed in this case). Replicates the functionality of Stata's cmcmmixlogit.

numMultiStarts
is the number of times to run the optimization loop, each time starting from a different random starting point for each parameter between startParBounds. Recommended for non-convex models, such as WTP space models and mixed logit models. Defaults to 1.

useAnalyticGrad
Set to FALSE to use numerically approximated gradients instead of analytic gradients during estimation. For now, using the analytic gradient is faster for MNL models but slower for MXL models. Defaults to TRUE.

scaleInputs
By default each variable in data is scaled to be between 0 and 1 before running the optimization routine because it usually helps with stability, especially if some of the variables have very large or very small values (e.g. > 10^3 or < 10^-3). Set to FALSE to turn this feature off. Defaults to TRUE.

startParBounds
sets the lower and upper bounds for the starting parameters for each optimization run, which are generated by runif(n, lower, upper). Defaults to c(-1, 1).

standardDraws
By default, a new set of standard normal draws are generated during each call to logitr (the same draws are used during each multistart iteration). The user can override those draws by providing a matrix of standard normal draws if desired. Defaults to NULL.

numDraws
The number of Halton draws to use for MXL models for the maximum simulated likelihood. Defaults to 50.

startVals
is vector of values to be used as starting values for the optimization. Only used for the first run if numMultiStarts > 1. Defaults to NULL.

options
A list of options for controlling the nloptr() optimization. Run nloptr::nloptr.print.options() for details.

Value

The function returns a list object containing the following objects.

Value	Description

coef	The model coefficients at convergence.

logLik	The log-likelihood value at convergence.

nullLogLik	The null log-likelihood value (if all coefficients are 0).

gradient	The gradient of the log-likelihood at convergence.

hessian	The hessian of the log-likelihood at convergence.

startPars	The starting values used.

multistartNumber	The multistart run number for this model.

multistartSummary	A summary of the log-likelihood values for each multistart run (if more than one multistart was used).

time	The user, system, and elapsed time to run the optimization.

iterations	The number of iterations until convergence.

message	A more informative message with the status of the optimization result.

status	An integer value with the status of the optimization (positive values are successes). Use statusCodes() for a detailed description.

call	The matched call to logitr().

inputs	A list of the original inputs to logitr().

data	A list of the original data provided to logitr() broken up into components used during model estimation.

numObs	The number of observations.

numParams	The number of model parameters.

freq	The frequency counts of each choice alternative.

modelType	The model type, 'mnl' for multinomial logit or 'mxl' for mixed logit.

weightsUsed	TRUE or FALSE for whether weights were used in the model.

numClusters	The number of clusters.

parSetup	A summary of the distributional assumptions on each model parameter ("f"="fixed", "n"="normal distribution", "ln"="log-normal distribution").

parIDs	A list identifying the indices of each parameter in coef by a variety of types.

scaleFactors	A vector of the scaling factors used to scale each coefficient during estimation.

standardDraws	The draws used during maximum simulated likelihood (for MXL models).

options	A list of options for controlling the nloptr() optimization. Run nloptr::nloptr.print.options() for details.

Details

The the options argument is used to control the detailed behavior of the optimization and must be passed as a list, e.g. options = list(...). Below are a list of the default options, but other options can be included. Run nloptr::nloptr.print.options() for more details.

Argument	Description	Default

xtol_rel	The relative x tolerance for the nloptr optimization loop.	1.0e-6

xtol_abs	The absolute x tolerance for the nloptr optimization loop.	1.0e-6

ftol_rel	The relative f tolerance for the nloptr optimization loop.	1.0e-6

ftol_abs	The absolute f tolerance for the nloptr optimization loop.	1.0e-6

maxeval	The maximum number of function evaluations for the nloptr optimization loop.	1000

algorithm	The optimization algorithm that nloptr uses.	"NLOPT_LD_LBFGS"

print_level	The print level of the nloptr optimization loop.	0

Examples

# For more detailed examples, visit
# https://jhelvy.github.io/logitr/articles/

library(logitr)

# Estimate a MNL model in the Preference space
mnl_pref <- logitr(
  data   = yogurt,
  choice = "choice",
  obsID  = "obsID",
  pars   = c("price", "feat", "brand")
)

#> Running Model...

#> Done!

# Estimate a MNL model in the WTP space, using a 10-run multistart
mnl_wtp <- logitr(
  data           = yogurt,
  choice         = "choice",
  obsID          = "obsID",
  pars           = c("feat", "brand"),
  price          = "price",
  modelSpace     = "wtp",
  numMultiStarts = 10
)

#> Running Multistart 1 of 10...

#> Running Multistart 2 of 10...

#> Running Multistart 3 of 10...

#> Running Multistart 4 of 10...

#> Running Multistart 5 of 10...

#> Running Multistart 6 of 10...

#> Running Multistart 7 of 10...

#> Running Multistart 8 of 10...

#> Running Multistart 9 of 10...

#> Running Multistart 10 of 10...

#> Done!

# Estimate a MXL model in the Preference space with "feat" and "brand"
# following normal distributions
mxl_pref <- logitr(
  data   = yogurt,
  choice = "choice",
  obsID  = "obsID",
  pars   = c("price", "feat", "brand"),
  randPars = c(feat = "n", brand = "n")
)

#> Running Model...

#> Done!