Simulate expected shares

Returns the expected shares for a set of one or more alternatives based on the results from an estimated model.

simulateShares(model, alts, obsIDName = NULL, priceName = NULL, alpha = 0.025)

Arguments

model
The output of a model estimated model using the logitr() function.

alts
A data frame of a set of alternatives for which to simulate shares. Each row is an alternative and each column an attribute corresponding to parameter names in the estimated model.

obsIDName
The name of the column that identifies each set of alternatives. Required if simulating results for more than one set of alternatives. Defaults to NULL (for a single set of alternatives).

priceName
The name of the parameter that identifies price. Only required for WTP space models. Defaults to NULL.

alpha
The sensitivity of the computed confidence interval, e.g. a 90% CI is obtained with alpha = 0.05. Defaults to alpha = 0.025.

Value

A data frame with the estimated shares for each alternative in alts.

Examples

# Run a MNL model in the Preference Space:
library(logitr)

mnl_pref <- logitr(
  data = yogurt,
  choiceName = "choice",
  obsIDName = "obsID",
  parNames = c("price", "feat", "dannon", "hiland", "yoplait")
)

#> Running Model...

#> Done!

# Create a set of alternatives for which to simulate shares:
alts <- subset(yogurt, obsID == 42,
  select = c("feat", "price", "dannon", "hiland", "yoplait")
)
row.names(alts) <- c("dannon", "hiland", "weight", "yoplait")

#> Warning: Setting row names on a tibble is deprecated.

alts

#> # A tibble: 4 x 5
#>    feat price dannon hiland yoplait
#> * <dbl> <dbl>  <dbl>  <dbl>   <dbl>
#> 1     0  6.30      1      0       0
#> 2     1  6.10      0      1       0
#> 3     0  7.90      0      0       0
#> 4     0 11.5       0      0       1

# Run the simulation using the estimated preference space MNL model:
simulateShares(mnl_pref, alts)

#>              share_mean  share_low share_high
#> Alt: dannon  0.60766441 0.54769558  0.6599274
#> Alt: hiland  0.02601869 0.01843736  0.0362452
#> Alt: weight  0.17802424 0.16328942  0.1922961
#> Alt: yoplait 0.18829266 0.13852862  0.2503708