Prints a summary of a model estimated using the 'logitr' package

Prints a summary of a model estimated using the 'logitr' package

# S3 method for logitr
summary(model)

Arguments

model
The output of a model estimated model using the logitr() function.

Value

Prints a summary of the model results to the console.

Examples

# Run a MNL model in the Preference Space with a multistart:
data(yogurt)

mnl_pref = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'),
  options    = list(
    numMultiStarts = 5,
    keepAllRuns    = TRUE))

#> Running Multistart 1 of 5 
#> Running Multistart 2 of 5 
#> Running Multistart 3 of 5 
#> Running Multistart 4 of 5 
#> Running Multistart 5 of 5 
#> Done!

# View a summary of the model:
summary(mnl_pref)

#> =================================================
#> SUMMARY OF ALL MULTISTART RUNS:
#> 
#>   run    logLik iterations status
#> 1   1 -2656.888         22      3
#> 2   2 -2656.888         20      3
#> 3   3 -2656.888         21      3
#> 4   4 -2656.888         23      3
#> 5   5 -2656.888         22      3
#> ---
#> To view meaning of status codes, use logitr.statusCodes() 
#> 
#> Summary of BEST model below (run with largest log-likelihood value)
#> =================================================
#> MODEL SUMMARY: 
#>                           
#> Model Space:    Preference
#> Model Run:          5 of 5
#> Iterations:             22
#> Elapsed Time:  0h:0m:0.11s
#> Weights Used?:       FALSE
#> 
#> Model Coefficients: 
#>          Estimate StdError    tStat pVal signif
#> price   -0.366584 0.024366 -15.0449    0    ***
#> feat     0.491433 0.120063   4.0931    0    ***
#> dannon   0.641184 0.054498  11.7652    0    ***
#> hiland  -3.074417 0.145384 -21.1469    0    ***
#> yoplait  1.375755 0.088982  15.4611    0    ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Model Fit Values: 
#>                                      
#> Log.Likelihood.         -2656.8878779
#> Null.Log.Likelihood.    -3343.7419990
#> AIC.                     5323.7758000
#> BIC.                     5352.7168000
#> McFadden.R2.                0.2054148
#> Adj..McFadden.R2            0.2039195
#> Number.of.Observations.  2412.0000000