Compare WTP from preference and WTP space models

Returns a comparison of the WTP between a preference space and WTP space model.

wtpCompare(model_pref, model_wtp, priceName)

Arguments

model_pref
The output of a "preference space" model estimated using the logitr() function.

model_wtp
The output of a "willingness to pay space" model estimated using the logitr() function.

priceName
The name of the parameter that identifies price.

Value

A data frame comparing the WTP estimates from preference space and WTP space models.

Details

Willingness to pay (WTP) is first computed from the preference space model by dividing the estimated parameters by the price parameter. Then those estimates are compared against the WTP values directly estimated from the "WTP" space model. Uncertainty is handled via simulation.

Examples

# Run a MNL model in the Preference Space:
library(logitr)

mnl_pref <- logitr(
  data = yogurt,
  choiceName = "choice",
  obsIDName = "obsID",
  parNames = c("price", "feat", "dannon", "hiland", "yoplait")
)

#> Running Model
#> Done!

# Get the WTP implied from the preference space model
wtp_mnl_pref <- wtp(mnl_pref, priceName = "price")

# Run a MNL model in the WTP Space:
mnl_wtp <- logitr(
  data = yogurt,
  choiceName = "choice",
  obsIDName = "obsID",
  parNames = c("feat", "dannon", "hiland", "yoplait"),
  priceName = "price",
  modelSpace = "wtp",
  options = list(startVals = wtp_mnl_pref$Estimate)
)

#> Running Model
#> **Using User Provided Starting Values For This Run**
#> Done!

# Compare the WTP between the two spaces:
wtpCompare(mnl_pref, mnl_wtp, priceName = "price")

#>                 pref           wtp difference
#> lambda      0.366584     0.3665834 -6.200e-07
#> feat        1.340568     1.3405856  1.757e-05
#> dannon      1.749079     1.7490883  9.330e-06
#> hiland     -8.386650    -8.3866858 -3.582e-05
#> yoplait     3.752905     3.7528902 -1.480e-05
#> logLik  -2656.887878 -2656.8878780 -1.000e-08