Compare WTP from preference and WTP space models

Returns a comparison of the WTP between a preference space and WTP space model.

wtpCompare(model_pref, model_wtp, price)

Arguments

model_pref
The output of a "preference space" model estimated using the logitr() function.

model_wtp
The output of a "willingness to pay space" model estimated using the logitr() function.

price
The name of the parameter that identifies price.

Value

A data frame comparing the WTP estimates from preference space and WTP space models.

Details

Willingness to pay (WTP) is first computed from the preference space model by dividing the estimated parameters by the price parameter. Then those estimates are compared against the WTP values directly estimated from the "WTP" space model. Uncertainty is handled via simulation.

Examples

library(logitr)

# Estimate a MNL model in the Preference space
mnl_pref <- logitr(
  data   = yogurt,
  choice = "choice",
  obsID  = "obsID",
  pars   = c("price", "feat", "brand")
)

#> Running Model...

#> Done!

# Compute the WTP implied from the preference space model
wtp_mnl_pref <- wtp(mnl_pref, price = "price")

# Estimate a MNL model in the WTP Space, using the computed WTP values
# from the preference space model as starting points
mnl_wtp <- logitr(
  data       = yogurt,
  choice     = "choice",
  obsID      = "obsID",
  pars       = c("feat", "brand"),
  price      = "price",
  modelSpace = "wtp",
  startVals  = wtp_mnl_pref$Estimate
)

#> Running Model...

#> NOTE: Using user-provided starting values for this run

#> Done!

# Compare the WTP between the two spaces
wtpCompare(mnl_pref, mnl_wtp, price = "price")

#>                       pref          wtp  difference
#> lambda           0.3665546     0.366589  0.00003448
#> feat             1.3406987     1.340815  0.00011655
#> brandhiland    -10.1362190   -10.135947  0.00027163
#> brandweight     -1.7490940    -1.749231 -0.00013707
#> brandyoplait     2.0038476     2.003530 -0.00031745
#> logLik       -2656.8878790 -2656.887881 -0.00000180